Add separate class for Kyber trading methods
